requirements.txt
tensorflow==2.7.0
tensorflow-datasets==4.4.0
Pillow==8.4.0
pandas==1.3.4
numpy==1.21.4
scipy==1.7.3
urllib3
matplotlib